# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mentis Semear API",
        "description": "API do monólito Mentis Semear (autenticação JWT, workshops, check-ins, avaliações e diário).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Mesma origem (dev e produção)"
        }
    ],
    "paths": {
...
        "/api/diaries": {
            "get": {
                "tags": [
                    "Diaries"
                ],
                "summary": "Lista diários (paginado)",
                "operationId": "f8edff380fdb2168574b18cff9182d7d",
                "parameters": [
                    {
                        "name": "workshop_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Diary"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Diaries"
                ],
                "summary": "Cria o diário de um workshop (1 por workshop)",
                "operationId": "c7845a0e85d0863872c88bc8cbf6a41b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "workshop_id",
                                    "title",
                                    "description",
                                    "datetime"
                                ],
                                "properties": {
                                    "workshop_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Relato do dia"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-09-01 16:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Diary"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "422": {
                        "description": "Falha de validação (workshop inexistente ou já com diário)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/diaries/{diary}": {
            "get": {
                "tags": [
                    "Diaries"
                ],
                "summary": "Detalha um diário",
                "operationId": "ea1b1515584fd0f83986a0c3630ce217",
                "parameters": [
                    {
                        "name": "diary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Diário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Diary"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Diaries"
                ],
                "summary": "Atualiza um diário (campos parciais)",
                "operationId": "7c6961f8ae6df4bbdf0e99fb0abfdd4f",
                "parameters": [
                    {
                        "name": "diary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "datetime": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Diary"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrado"
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Diaries"
                ],
                "summary": "Remove um diário (evidências são removidas em cascata)",
                "operationId": "948da250b19d1202c3aa8235334863ea",
                "parameters": [
                    {
                        "name": "diary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removido"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrado"
                    },
                    "409": {
                        "description": "Vínculo impede a remoção"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Diary": {
                "properties": {
                    "id": {
                        "description": "OpenAPI schema only — never instantiated.",
                        "type": "integer",
                        "example": 1
                    },
                    "workshop_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_creator_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "example": "Relato do dia"
                    },
                    "description": {
                        "type": "string",
                        "example": "Como foi a dinâmica..."
                    },
                    "datetime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envie o token no header: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}